% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagpipe.R
\name{bagpipe}
\alias{bagpipe}
\title{Identify candidate alpha/beta pairs.}
\usage{
bagpipe(alpha, beta, replicates = 100, frac = 0.75, thres = c(0.3, 0.6,
  0.9), bootstrap = FALSE)
}
\arguments{
\item{alpha}{Matrix recording which alpha chains appear in each well of the
data. See  .}

\item{beta}{Matrix recording which beta chains appear in the each well of the
data. See .}

\item{replicates}{The number of times the resampling procedure is repeated, i.e.
the number of replicates. At least 100 replicates is recommended.}

\item{frac}{The fraction of the wells resampled in each replicate. Default
is 75\% of the wells}

\item{thres}{A vector of thresholds (from 0 to 1) to use to filter the
results from all of the candidate alpha/beta pairs from the resampling
procedure.}

\item{bootstrap}{Legacy option. Calls a bootstrapping strategy (which
resamples with replacement) instead of sampling a subset without
replacement.}
}
\value{
A n x 2 matrix where
}
\description{
\code{bagpipe()} uses the alphabetr resampling procedure on sequencing data
   to identify candidate alpha/beta pairs. The procedure takes a subsample
   of the data without replacement, calculates association scores using ASOIDHAOISJ,
   and then for each well uses the Hungarian algorithm to determine the most
   likely pairings for the chains found in the well. Each time this is done
   is a replicate, and the number of replicates is specified as an option.
   A threshold is then used to filter the candidate pairs that appear in
   proportion of the replicates larger than the threshold, resulting in the
   final list of candidate pairs. Bagpipe is an acronym for
   \strong{b}ootstrapping \strong{a}lphabetr \strong{g}enerated
   \strong{p}a\strong{i}rs \strong{p}rocedur\strong{e} (based on older
   versions that utilized bootstrapping)
}
\examples{
set.seed(110)
createClones()
createData()
bagpipe(alpha = data_alph, beta = data_beta)
}

