% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_scores.R
\name{chain_scores}
\alias{chain_scores}
\title{Calculate association scores between alpha and beta chain pairs.}
\usage{
chain_scores(data_a, data_b)
}
\arguments{
\item{data_a}{Matrix recording which alpha chains appear in each well of the
data. See .}

\item{beta}{Matrix recording which beta chains appear in the each well of the
data. See .}

\item{numb_wells}{Vector containing the number of cells sampled in the wells
of each column of the plates.}
}
\value{
A list containing the alpha and beta association scores. Accessed
   with list$alpha and list$beta respectively.
}
\description{
\code{chain_scores()} calculates association scores between every pair of
   alpha and beta chains based on the number of concurrent well appearances
   each alpha and beta pair makes, scaled inversely by the number of unique
   chains in that well. See Lee et. al. 2016 for more information about this
   procedure.
}
\examples{
   set.seed(110)
   createClones()
   createData()
   scores <- chain_scores(data_a = data_alph, data_b = data_beta,
                          numb_cells = cells_vec)
   scores <- scores$alpha + scores$beta
}

