% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clones.R
\name{create_clones}
\alias{create_clones}
\title{Create a synthetic list of clones}
\usage{
create_clones(numb_beta = 1000, dual = 0.1)
}
\arguments{
\item{numb_beta}{The number of unique betas in the population}

\item{dual}{The proportion of clones that are dual TCR clones, i.e. has two
distinct alpha chains}
}
\description{
\code{create_clones()} creates a list of (beta, alpha1, alpha2) sets that
   represents the indices of the chains of clones. The function will take a
   fixed number of unique beta chains that are in the T cell population, and
   then use the degree of beta and alpha sharing to determine the number of
   unique alpha chains in the populations. These chains will then be randomly
   assigned to each other, with a proportion of them being dual TCR clones
   (i.e. alpha1 != alpha2), forming our random list of clones with their
   chain indices.
}

