% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_clones_tail.R
\name{dual_clones_tail}
\alias{dual_clones_tail}
\title{Discriminate between beta sharing clones and dual TCR clones.}
\usage{
dual_clones_tail(alpha, beta, freq_results, cells, population)
}
\arguments{
\item{freq_results}{Output of \code{freq_estimate()}}

\item{data_alpha}{Matrix recording which alpha chains appear in each well of the
data. See  .}

\item{data_beta}{Matrix recording which beta chains appear in the each well of the
data. See .}

\item{numb_wells}{Vector containing the number of cells sampled in the wells
of each column of the plates.}
}
\value{
A n x 3 matrix where n is the number of candidate clones, column 1
   is the beta index of the clone, and column 2-3 are the alpha indices of
   the clone
}
\description{
\code{dual_clones()} distinguishes between clones that share a common beta
   chain and dual TCR clones with two productive alpha chains. The procedure
   tests the null hypothesis that two candidate alpha, beta pairs with the
   same beta represent two separate clones by using the frequency estimates
   to calculate the number of wells that both clones are expected to be in.
   This is compared to the actual number of wells that both clones appear in,
   and if the actual number is greater than the expected number, than the
   pairs are chosen to represent a dual TCR clone.
}
\examples{
asdfas
}

