% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dual_discrim_dual_likelihood}
\alias{dual_discrim_dual_likelihood}
\title{Calculate likelihood of dual clone}
\usage{
dual_discrim_dual_likelihood(est, err, numb_cells, numb_wells, binomials)
}
\arguments{
\item{est}{a}

\item{err}{a}

\item{numb_cells}{a}

\item{numb_wells}{a}

\item{binomials}{a}
}
\description{
Calculate likelihood of dual clone
}

