% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_top.R
\name{dual_top}
\alias{dual_top}
\title{Discriminate between beta-sharing clones and dual-alpha TCR clones (optimized
for common clones)}
\usage{
dual_top(alpha, beta, pair, error, cells)
}
\description{
\code{dual_tail()} distinguishes between clones that share a common beta
   chain and dual TCR clones with two productive alpha chains. The procedure
   tests the null hypothesis that two candidate alpha, beta pairs with the
   same beta represent two separate clones by using the frequency estimates
   to calculate the number of wells that both clones are expected to be in.
   This is compared to the actual number of wells that both clones appear in,
   and if the actual number is greater than the expected number, than the
   pairs are chosen to represent a dual TCR clone.
}

