% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_estimate.R
\name{freq_estimate}
\alias{freq_estimate}
\title{Estimation frequencies of clones identified by alphabetr}
\usage{
freq_estimate(alpha, beta, pair, error = 0.15, cells)
}
\arguments{
\item{alpha}{Matrix recording which alpha chains appear in each well of the
data. See  .}

\item{beta}{Matrix recording which beta chains appear in the each well of the
data. See .}

\item{pair}{A matrix where each row is a beta/alpha pair, column 1 is the
beta index, and column 2 is the alpha index.}

\item{error}{The error or "dropped" chain rate due to PCR or sequencing
errors in the experimental pipeline.}

\item{cells}{The number of cells per well in each column of the plates.
Should be a vector of 12 elements.}
}
\value{
A data frame with frequency estimates and confidence intervals
}
\description{
\code{freq_estimate()} estimates the frequencies of clones with confidence
   intervals by using a maximum likelihood approach. The function looks at
   the wells that a chains of a clone appear in and determines the most
   likely frequency that explains the data.
}
\examples{
TCR <- create_clones()
data <- create_data()
results <- bagpipe(alpha = data$alpha, beta = data$beta, )
freq <- freq_estimate(data$alpha, data$beta, results, 0.15, cellstor)
}

