% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihood_dual}
\alias{likelihood_dual}
\title{Calculate likelihood of a single TCR clone}
\usage{
likelihood_dual(est, err, numb_wells, numb_cells, numb_sample)
}
\arguments{
\item{est}{a}

\item{err}{a}

\item{numb_wells}{a}

\item{numb_cells}{a}

\item{numb_sample}{a}
}
\description{
Calculate likelihood of a single TCR clone
}

