% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihood_single}
\alias{likelihood_single}
\title{Calculate likelihood curve of frequency estimates for a single TCR clone}
\usage{
likelihood_single(est, err, numb_wells, numb_cells, numb_sample)
}
\arguments{
\item{est}{Clonal frequency estimate}

\item{err}{Mean drop error rate}

\item{numb_wells}{A vector with the number of wells with the distinct
sample sizes}

\item{numb_cells}{A vector of the distinct sample sizes, i.e. the number of
cells per well}

\item{numb_sample}{A vector with the number of wells of the sample size of
the same position of \code{numb_cells} that contains the alpha and beta
chains the clone}
}
\value{
A numeric with the negative log likelihood
}
\description{
Calculate likelihood curve of frequency estimates for a single TCR clone
}

