% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dual_discrim_shared_likelihood}
\alias{dual_discrim_shared_likelihood}
\title{Calculate likelihood of dual clone}
\usage{
dual_discrim_shared_likelihood(est1, est2, err, numb_cells, numb_wells,
  binomials, multinomials)
}
\arguments{
\item{err}{a}

\item{numb_cells}{a}

\item{numb_wells}{a}

\item{binomials}{a}

\item{multinomials}{a}

\item{est}{a}
}
\description{
Calculate likelihood of dual clone
}

